requirements.txt
fastapi
uvicorn[standard]
Jinja2
python-multipart
python-dotenv
google-genai
pillow
requests
pydantic
typing_extensions
